Update singleRound(), Add game(), update code
rename singleRound() with playRound() and add console.log() to display
the computer & player selection in the func, update parameters to receive
values at the time of func call and add the condition for null value of
player selection. Add 5 round play code to game() func. Update the
counter comparison code to add logic for equality. Make changes in
comments.